Move interceptors to a separate directory
